♻️ Refactor : 검색 페이지 blocks 폴더 구조 수정
